Move the symbol package into the compiler package.
