+++ internal/vm/compiler/symbol/symtab.go
