+++ internal/vm/compiler/symbol/symtab_test.go
